Renamed top tokens balance hook
